v18.9.30
*All controllers implemented.

*Added Lazy Loading Proxies to allow Entity Framework to load related entities and return it.

*Added an Update method to the services.

+++ MyGamesList API/Repository/GameRepository.cs
@@ -16,21 +16,44 @@ namespace MyGamesListAPI.Repository {
 
         public IEnumerable<Game> GetGames => db.Games;
 
-        public void Add(Game game) {
-            db.Games.Add(game);
-            db.SaveChanges();
-        }
-
         public Game GetGame(long id) {
             Game game = db.Games.Find(id);
             return game;
         }
 
-        public void Remove(long id) {
-            Game game = db.Games.Find(id);
+        public void Add(Game game) {
+            db.Games.Add(game);
+            db.SaveChanges();
+        }
+
+        public bool Remove(long id) {
+            var game = db.Games.Find(id);
+
+            if (game == null) {
+                return false;
+            }
+
             db.Games.Remove(game);
             db.SaveChanges();
+            return true;
+        }
 
+        public bool Update(long id, Game newGame) {
+            var game = db.Games.Find(id);
+
+            if(game == null) {
+                return false;
+            }
+
+            game.SteamAppid = newGame.SteamAppid;
+            game.Title = newGame.Title;
+            game.ShortDescription = newGame.ShortDescription;
+            game.DetailedDescription = newGame.DetailedDescription;
+            game.HeaderImage = newGame.HeaderImage;
+            game.Price = newGame.Price;
+
+            db.SaveChanges();
+            return true;
         }
     }
 }
