+++ MyGamesList API/Repository/WishlistItemRepository.cs
@@ -23,11 +23,34 @@ namespace MyGamesListAPI.Repository {
             return db.WishlistItems.Find(id);
         }
 
-        public void Remove(long id) {
+        public bool Remove(long id) {
             WishlistItem wishlistItem = db.WishlistItems.Find(id);
+
+            if (wishlistItem == null) {
+                return false;
+            }
 
             db.WishlistItems.Remove(wishlistItem);
             db.SaveChanges();
+            return true;
+        }
+
+        public bool Update(long id, WishlistItem newItem) {
+            var item = db.WishlistItems.Find(id);
+            
+            if (item == null) {
+                return false;
+            }
+
+            if (item.Game.Id != newItem.Game.Id) {
+                var game = db.Games.Find(newItem.Game.Id);
+                item.Game = game;
+            }
+
+            item.RankPosition = newItem.RankPosition;
+
+            db.SaveChanges();
+            return true;
         }
     }
 }
